# backend/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple Kanban simulation API.",
        "title": "Kanban-Sim API",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "Peter Kerschbaumer",
            "email": "you@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/games": {
            "get": {
                "description": "Returns a list of all games in the system.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "List all games",
                "responses": {
                    "200": {
                        "description": "List of games",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.Game"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new game using the embedded default board; no request body required",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Create a new game",
                "responses": {
                    "201": {
                        "description": "New game created",
                        "schema": {
                            "$ref": "#/definitions/response.CreateGameResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/games/{id}": {
            "get": {
                "description": "Returns the full game record for the given UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Get game by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/response.GameResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid or missing game ID",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Game not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Removes the game record identified by the given UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Delete game by ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid or missing game ID",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Game not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates the specified game’s current day by its UUID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Update game day",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New game day",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updateGameRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid game ID or JSON payload",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Game not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/games/{id}/columns": {
            "get": {
                "description": "Returns the list of columns (including subcolumns) belonging to the specified game UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "columns"
                ],
                "summary": "List columns by game ID",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of columns",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Column"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid or missing game ID",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Game not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.updateGameRequest": {
            "type": "object",
            "properties": {
                "day": {
                    "type": "integer"
                }
            }
        },
        "models.Column": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "orderIndex": {
                    "type": "integer"
                },
                "parentId": {
                    "description": "nil for top-level",
                    "type": "string"
                },
                "subColumns": {
                    "description": "built in memory",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Column"
                    }
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "description": "\"active\", queue\", \"done\"",
                    "type": "string"
                },
                "wipLimit": {
                    "description": "only set if non-zero",
                    "type": "integer"
                }
            }
        },
...
        "models.Game": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "day": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "response.CreateGameData": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "7d7881cf-8d9f-457f-ac93-aa498ea8c0af"
                }
            }
        },
        "response.CreateGameResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/response.CreateGameData"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "contextual error message"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "response.GameResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Game"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    }
}